Adjust operate switch statement default
Previously, division was used as the default operation and would be performed if a user pressed equals before pressing an operator button. This would typically reset the display to 0 (the first number in the operation wouldn't have been set and would remain 0), although it could result in a division by 0 error. By switching the default operation to addition, pressing equals before pressing an operator button simply keeps the currently displayed number on the screen.